+++ lib/openapi_first/response_validation.rb
@@ -1,12 +1,12 @@
 # frozen_string_literal: true
 
 require 'multi_json'
-require_relative 'response_validation/middleware'
+require_relative 'middlewares/response_validation'
 
 module OpenapiFirst
   module ResponseValidation
     def self.new(app, options = {})
-      Middleware.new(app, options)
+      OpenapiFirst::Middlewares::ResponseValidation.new(app, options)
     end
   end
 end